Fix for xisoslice, many additions
Important fix to `xisoslice` which was causing a choke point.

This push has a bunch of rearranging within the code. This was mostly to be more
organized by moving all attribute-setting outside of the accessor so that functions
return the same results whether they are run with the accessor or not. All
functionality is now available either way.

Also added a bunch of "basic" calculations which are wrappers for xarray
functions that allow for changing grid and maintain attributes. Also grid-based
mean and sum (gridmean, gridsum). Also a wrapper for downsampling in time.