Lots of updates, esp documentation.
The form of xroms is settling down now. The package has largely but not fully
integrated the use of cf-xarray now, which has simplified some of the code.
Using cf-xarray also removes the need for some of the wrappers used earlier.
The docs have been standardized so that readthedocs can be used in the future.
Still need to do docs for the xroms.py file. Still need to update the
interpolation functionality and notebook.