Dataset being read in can have a single layer
Previously a dataset will error out when read in if it doesn't have
certain variables that are being checked for. This particularly
happens in `xroms.roms_dataset` where variables associated with
calculating the depths are assumed to be present, making it
impossible to read in a dataset without that depth information.

This commit searches for the presence of the depth variable `s_rho` in
the dataset and if present checks it length to be more than 1. If
both of these are true, a new variable `ds['3d']` is set to True. If
not, `ds['3d']=False`. This variable is checked several times to
determine whether or not to use the variables needed to calculate
depth-related variables subsequently, avoiding the previous issues.