Update README.md
I ran into some problems installing xroms with the directions in the README.md. It turns out that Github updated their security protocols a while back and so the installation guide needs to be updated. I made some very minor changes to the README.md in order for the installation be successful using the given instructions.